requirements.txt
helpers==0.2.0
matplotlib==3.10.8
networkx==3.3
numpy>=1.23,<2.0
pandas>=2.3,<2.4
PyBioMed==1.0
PyYAML==6.0.3
rdkit==2025.3.3
requests==2.32.5
scikit-learn
scipy>=1.10,<1.16
torch>=2.5
torch-geometric>=2.7
tqdm==4.67.1
